+++ R/utils.R
@@ -8,6 +8,9 @@
 #'
 #' @return A `data.frame`.
 #'
+#' @details `read_logs()` returns a `data.frame` with the empty character columns "timestamp", "log_lvl" and "log_msg"
+#' if the log file has no entries.
+#'
 #' @examples
 #'   set_logfile(file.path(tempdir(), "loggit.log"), confirm = FALSE)
 #'   message("Test log message")
@@ -18,7 +21,11 @@ read_logs <- function(logfile = get_logfile(), unsanitizer = default_ndjson_unsa
 
   base::stopifnot("Log file does not exist" = file.exists(logfile))
 
-  read_ndjson(logfile, unsanitizer = unsanitizer)
+  log <- read_ndjson(logfile, unsanitizer = unsanitizer)
+
+  if(nrow(log) == 0L) log <- data.frame(timestamp = character(), log_lvl = character(), log_msg = character())
+
+  return(log)
 }